+++ sorellaFront/src/app/application/services/tests/auth.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { AuthService } from './auth.service';
+import { AuthService } from '../auth.service';
 
 describe('AuthService', () => {
   let service: AuthService;
