+++ sorellaFront/src/app/application/services/tests/user-service.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { UserServiceService } from './user-service.service';
+import { UserServiceService } from '../user-service.service';
 
 describe('UserServiceService', () => {
   let service: UserServiceService;
